pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dieprompterei"
version = "0.1.0"
description = "Generic prompt optimization with sacred train/test separation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
    "click>=8.1.0",
    "scikit-learn>=1.3.0",
    "nltk>=3.8",
    "rich>=13.0.0",
]

[project.scripts]
prompterei = "dieprompterei.cli:cli"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.uv.workspace]
members = []
